requirements.txt
flask
flask-cors
smtplib
email-validator
python_dotenv
pymongo